Ensure expansion does not strip spaces
In substitution and elsewhere we want to conserve spaces.
Their number may not be significant in shell in general, but their
presence cannot be ignored, in particular in string substitutions.

To support this there is a new tokenize() functions that takes care
of the shell lexing keeping spaces and grouping them together.

In addition:

- add new strict argument to raise exception if a parameter is missing
  from the environment
- add optional debug logging/tracing and several space-related tests.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>